Refactor command handling into its own package and split catfact into separate function
